checkout to main

+++ src/main/java/lms/service/impl/TrashServiceImpl.java
@@ -6,7 +6,6 @@ import lms.dto.response.TrashResponse;
 import lms.entities.Trash;
 import lms.entities.User;
 import lms.enums.Role;
-import lms.repository.CourseRepository;
 import lms.repository.StudentRepository;
 import lms.repository.TrashRepository;
 import lms.repository.UserRepository;
@@ -14,7 +13,6 @@ import lms.service.TrashService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
-import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
@@ -25,23 +23,23 @@ public class TrashServiceImpl implements TrashService {
 
     private final TrashRepository trashRepository;
     private final UserRepository userRepository;
-    private final CourseRepository courseRepository;
-    private final StudentRepository studentRepository;
     private final TrashDeleteService trashDeleteService;
+    private final StudentRepository studentRepository;
 
     @Override
     public AllTrashResponse findAll(int page, int size) {
         User authUser = userRepository.getByEmail(SecurityContextHolder.getContext().getAuthentication().getName());
-        Pageable pageable = PageRequest.of(page - 1, size);
         if (authUser.getRole().equals(Role.ADMIN)) {
-            Page<TrashResponse> allTrash = trashRepository.findAllTrash(pageable);
+            Page<TrashResponse> instructorTrash =
+                    trashRepository.findAllTrash(PageRequest.of(page - 1, size));
             return AllTrashResponse.builder()
-                    .page(allTrash.getNumber() + 1)
-                    .size(allTrash.getNumberOfElements())
-                    .trashResponses(allTrash.getContent())
+                    .page(instructorTrash.getNumber() + 1)
+                    .size(instructorTrash.getNumberOfElements())
+                    .trashResponses(instructorTrash.getContent())
                     .build();
         }
-        Page<TrashResponse> instructorTrash = trashRepository.findAllTrashByInstructorId(authUser.getId(), pageable);
+        Page<TrashResponse> instructorTrash =
+                trashRepository.findAllTrashByAuthId(authUser.getId(), PageRequest.of(page - 1, size));
         return AllTrashResponse.builder()
                 .page(instructorTrash.getNumber() + 1)
                 .size(instructorTrash.getNumberOfElements())
@@ -53,29 +51,31 @@ public class TrashServiceImpl implements TrashService {
     @Override
     public SimpleResponse restoreData(Long trashId) {
         Trash trash = trashRepository.findByIdOrThrow(trashId);
-        return null;
+        deleteTrash(trash, true);
+        trashRepository.deleteTrash(trashId);
+      return SimpleResponse.builder().httpStatus(HttpStatus.OK).message("Данные успешно восстановлены").build();
     }
 
     @Override
     public SimpleResponse delete(Long trashID) {
         Trash trash = trashRepository.findByIdOrThrow(trashID);
-        deleteTrash(trash);
+        deleteTrash(trash, false);
         return SimpleResponse.builder()
                 .httpStatus(HttpStatus.OK)
                 .message("Trash with id: %d deleted".formatted(trash.getId()))
                 .build();
     }
 
-    private void deleteTrash(Trash trash) {
+    private void deleteTrash(Trash trash, boolean isRestored) {
         switch (trash.getType()) {
-            case COURSE -> {
-                courseRepository.delete(trash.getCourse());
-                trashRepository.delete(trash);
-            }
+            case COURSE -> trashDeleteService.deleteGroup(trash);
             case STUDENT -> {
-                System.err.println("trash.getId() = " + trash.getId());
-                trashDeleteService.deleteStudent(trash.getStudent());
-                trashRepository.delete(trash);
+                if (isRestored) {
+                    studentRepository.getStudentByTrashId(trash.getId())
+                    trashRepository.delete(trash);
+                }
+                            trashDeleteService.deleteStudent(trash);
+                }
             }
         }
     }
